extracted direction logic to new service
